input/requirements.txt
evdev==1.4.0
nats-py==2.0.0rc5
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
